This is a larger commit, that represents (besides my bad consciousness) three different major changes:
1) Implementation of a metadata specification type: content types
2) Storage path handling and allowing for multiple storage locations
3) Observer handling

Because the changes span multiple files and each file implements two
or more of the above changes, this commit entails quite a lot of
files.
Here follows a list of classes and the changes made to them.
---

RepositoryObserver
RepositoryServer
JMSService
RESTObjectManagement
RESTRepositoryRelayer
RMIObjectManagement
RMIRepositoryServer
DefaultDigitalObject

Implemented content types that specify, for a given server
implementation, what type of object it recognizes. This information is
used in the notification of observers so that they are able to handle
content without prior examination.

---

RepositoryServer
JMSService
RESTRepositoryRelayer
RMIRepositoryServer
RMIServer
DBStorage
FileStorage
StorageProvider
DefaultDigitalObject
DefaultIdentifier
DigitalObject
ObjectIdentifier
RepositoryServerTest

Moved the responsibility of defining the storage paths for the objects
into the server implementations, making it easier to define various
storage paths differentiating on the object types

---

RepositoryServer
RMIRepositoryServer
RESTRepositoryRelayer

Moved the handling of observers out from the core server into the
server implementations